requirements.txt
aniso8601==8.1.0
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
config==0.5.1
dnspython==1.16.0
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
peewee==3.14.4
pycodestyle==2.7.0
PyJWT==2.3.0
pymongo==3.12.0
python-decouple==3.6
python-dotenv==0.20.0
pytz==2020.4
requests==2.27.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
twilio==7.8.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==1.0.1
zipp==3.8.0
